+++ example/data_source_templates.tf
@@ -1,0 +1,6 @@
+data "clouddk_templates" "example" {}
+
+output "data_clouddk_templates_example_result" {
+  description = "The templates"
+  value       = "${data.clouddk_templates.example.result}"
+}