+++ types.go
@@ -14,3 +14,12 @@ type LocationBody struct {
 
 // LocationListBody describes a datacenter location list payload.
 type LocationListBody []LocationBody
+
+// TemplateBody describes a datacenter location payload.
+type TemplateBody struct {
+	Identifier string `json:"identifier"`
+	Name       string `json:"name"`
+}
+
+// TemplateListBody describes a datacenter location list payload.
+type TemplateListBody []TemplateBody