+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 FEATURES:
 
 * **New Data Source:** `clouddk_locations`
+* **New Data Source:** `clouddk_templates`